tokens now updating on site, just need to also decrease in database
